Move idea list related types and functions to the Snippet module.
